pipeline/requirements.txt
numpy<2
feedparser==6.0.11
httpx==0.27.2
sentence-transformers==3.3.1
chromadb==0.5.23
groq==0.12.0
python-dotenv==1.0.1
datasets==3.2.0
ragas==0.2.6
langchain-groq==0.2.3
langchain-huggingface
